fixed some errors

+++ EmoloyerSystem/EmoloyerSystem/src/main/java/com/example/EmoloyerSystem/Repository/ApplicationRepository.java
@@ -4,11 +4,10 @@ import com.example.EmoloyerSystem.Entity.Application;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface ApplicationRepository extends JpaRepository<Application,Long> {
+public interface ApplicationRepository extends JpaRepository<Application,Integer> {
 
-    Optional<Application>findByID(Long ID);
+    Optional<Application>findByID(int ID);
 }
